requirements.txt
streamlit
opencv-python
numpy
scipy
pillow
pytesseract
